project_root/app/requirements.txt
fastapi
uvicorn
langchain
langchain-community
openai
sqlalchemy
python-dotenv
pyodbc